RePar: Move handle_content_length_header to parsing file
Where it is actually used, rather than in the mostly unrelated
Connection.cpp
