+++ src/ConnectionManager/Connection.cpp
@@ -11,7 +11,6 @@
 #include "Request.hpp"
 #include "Response.hpp"
 #include "config.hpp"
-#include "request_parser.hpp"
 
 Connection::Connection(const ConfigServer* const config, int socket)
     : _config(config),
@@ -86,28 +85,6 @@ ssize_t Connection::receive_data() {
     }
 
     return total;
-}
-
-bool Connection::handle_content_length_header(Request& request) {
-    size_t content_length = 0;
-
-    if (request.has_header("content-length") &&
-        !parse_content_length_value(request.header("content-length")->second, content_length)) {
-        request.set_error_status(400);
-        request.set_status(REQ_ERROR);
-        return false;
-    }
-    request.set_content_length(content_length);
-    if (content_length > _config->client_max_body_size) {
-        request.set_error_status(413);
-        request.set_status(REQ_ERROR);
-        return false;
-    }
-    if (content_length > 0) {
-        request.set_status(REQ_HEADERS);
-        return false;
-    }
-    return true;
 }
 
 bool Connection::has_pending_cgi() const {
